# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: BlogV2 Backend API
  description: |
    一个现代化的博客后端API，提供用户管理、文章发布、分类标签等功能。
    使用JWT令牌进行身份验证，支持基于角色的权限控制。
  version: 1.0.0
  contact:
    name: API 支持
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: 本地开发服务器
  - url: https://api.example.com/api/v1
    description: 生产服务器
...
paths:
...
  /users/me:
    get:
      tags:
        - 用户
      summary: 获取当前用户信息
      description: 获取当前已认证用户的详细信息
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功返回当前用户信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUser'
        '401':
          description: 未提供有效令牌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    patch:
      tags:
        - 用户
      summary: 更新当前用户信息
      description: 更新当前已认证用户的个人信息
      operationId: updateCurrentUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: 用户信息更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUser'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /users:
    get:
      tags:
        - 用户
      summary: 获取用户列表（分页）
      description: 获取用户列表，支持分页、排序和筛选。需要管理员权限。
      operationId: getUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: 页码（默认1）
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          description: 每页数量（默认10）
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: sortBy
          in: query
          description: 排序字段
          required: false
          schema:
            type: string
            enum: [createdAt, updatedAt, username, email]
        - name: sortOrder
          in: query
          description: 排序顺序
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: 成功返回用户列表和分页信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponseUser'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足（非管理员）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /users/{id}:
    get:
      tags:
        - 用户
      summary: 根据ID获取用户信息
      description: 根据用户ID获取用户详细信息。管理员可以查看任何用户，普通用户只能查看自己的信息。
      operationId: getUserById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 用户ID
          schema:
            type: string
      responses:
        '200':
          description: 成功返回用户信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUser'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 没有权限查看此用户信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    patch:
      tags:
        - 用户
      summary: 更新用户信息（管理员）
      description: 管理员更新指定用户的信息
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 用户ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: 用户信息更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUser'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足（非管理员）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    delete:
      tags:
        - 用户
      summary: 删除用户
      description: 删除指定用户账户。需要管理员权限。
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 用户ID
          schema:
            type: string
      responses:
        '200':
          description: 用户删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMessage'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足（非管理员）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  # ==================== 文章相关接口 ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT令牌认证
...
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 请求是否成功
        data:
          type: object
          description: 响应数据，具体类型取决于接口
          nullable: true
        error:
          $ref: '#/components/schemas/ApiError'
          nullable: true
        message:
          type: string
          description: 人类可读的消息
          nullable: true
        timestamp:
          type: string
          format: date-time
          description: 响应时间戳
      required:
        - success
        - data
        - error
        - message
        - timestamp
    ApiError:
      type: object
      properties:
        code:
          type: string
          description: 错误代码
        message:
          type: string
          description: 错误描述
        details:
          type: object
          description: 错误详情（可选）
          nullable: true
      required:
        - code
        - message
    ApiResponseMessage:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                message:
                  type: string
              required:
                - message

    # ===== 用户相关 =====
    User:
      type: object
      properties:
        id:
          type: string
          description: 用户唯一标识符
        username:
          type: string
          description: 用户名
        email:
          type: string
          format: email
          description: 电子邮箱
        role:
          type: string
          enum: [USER, EDITOR, ADMIN]
          description: 用户角色
        avatar:
          type: string
          description: 头像URL
          nullable: true
        bio:
          type: string
          description: 个人简介
          nullable: true
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
      required:
        - id
        - username
        - email
        - role
        - createdAt
        - updatedAt
...
    UpdateUserDto:
      type: object
      properties:
        username:
          type: string
          description: 用户名（3-20个字符）
          minLength: 3
          maxLength: 20
          nullable: true
        email:
          type: string
          format: email
          description: 电子邮箱
          nullable: true
        password:
          type: string
          description: 密码（至少8个字符）
          minLength: 8
          nullable: true
        avatar:
          type: string
          description: 头像URL
          nullable: true
        bio:
          type: string
          description: 个人简介
          nullable: true
        role:
          type: string
          enum: [USER, EDITOR, ADMIN]
          description: 用户角色
          nullable: true
...
    ApiResponseUser:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/User'
    PaginatedResponseUser:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/User'
            pagination:
              $ref: '#/components/schemas/PaginationInfo'

    # ===== 文章相关 =====
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          description: 当前页码
        pageSize:
          type: integer
          description: 每页数量
        totalItems:
          type: integer
          description: 总项目数
        totalPages:
          type: integer
          description: 总页数
        hasNextPage:
          type: boolean
          description: 是否有下一页
        hasPreviousPage:
          type: boolean
          description: 是否有上一页
      required:
        - page
        - pageSize
        - totalItems
        - totalPages
        - hasNextPage
        - hasPreviousPage

    # ===== 健康检查响应 =====
...